backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic-settings==2.7.1
python-multipart==0.0.20
facenet-pytorch>=2.6.0
Pillow>=10.0.0
numpy>=1.24.0
torch>=2.0.0
